added basic Dockerfile
