+++ units/.dockerignore
@@ -1,0 +1,6 @@
+*
+!unittab
+!units.c
+!inputs
+!outputs
+!test.sh
